chore: Remove service_name from ModelDeploymentCard (#2349)
