+++ lib/llm/src/local_model.rs
@@ -251,12 +251,15 @@ impl LocalModel {
         &self.full_path
     }
 
+    /// Human friendly model name. This is the correct name.
     pub fn display_name(&self) -> &str {
         &self.card.display_name
     }
 
+    /// The name under which we make this model available over HTTP.
+    /// A slugified version of the model's name, for use in NATS, etcd, etc.
     pub fn service_name(&self) -> &str {
-        &self.card.service_name
+        self.card.slug().as_ref()
     }
 
     pub fn request_template(&self) -> Option<RequestTemplate> {
