+++ lib/llm/src/preprocessor/prompt/template.rs
@@ -1,24 +1,12 @@
 // SPDX-FileCopyrightText: Copyright (c) 2024-2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
 
 use std::{collections::HashSet, sync::Arc};
 
 use anyhow::{Context, Ok, Result};
 use minijinja::Environment;
 
-use crate::model_card::model::{ModelDeploymentCard, PromptContextMixin, PromptFormatterArtifact};
+use crate::model_card::{ModelDeploymentCard, PromptContextMixin, PromptFormatterArtifact};
 
 mod context;
 mod formatters;
